implement article editor component
