+++ src/common/containers/news/article-editor/actions/index.jsx
