+++ src/common/containers/news/article-editor/config.js
