+++ src/common/containers/news/article-editor/fields/styles.scss
