+++ src/common/containers/news/article-editor/styles.scss
@@ -1,7 +1,7 @@
 @import "src/scss/mixins";
 @import "src/scss/variables";
 
-.create-news-page-container {
+.article-editor-container {
   width: 100%;
 
   gap: 40px;
