+++ src/common/containers/news/index.js
@@ -1,3 +1,4 @@
 export { NewsGrid } from "./news-grid";
 export { NewsAuthorInfo } from "./author-info";
 export { NewsArticleInfo } from "./article-info";
+export { ArticleEditor } from "./article-editor";
