--- src/screens/news/create-news/news-builder/styles.scss
@@ -1,13 +1,0 @@
-.news-builder-container {
-  gap: 24px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.news-actions-container {
-  gap: 48px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
