+++ src/screens/news/index.js
@@ -1,3 +1,4 @@
 export { ArticlePage } from "./article";
 export { NewsListPage } from "./news-list";
-export { CreateNewsPage } from "./create-news";
+export { EditArticlePage } from "./edit-article";
+export { CreateArticlePage } from "./create-article";
